fix Tor test
tor-test.js was assigning a javascript object to
EnigmailGpg.agentPath, rather than just a raw path.

This caused autocrypt-test.js to break, because EnigmailGpg.agentPath
is supposed to just be a simple string.

Fixing tor-test.js, and fixing tor.jsm to do the right thing, makes
the rest of the test suite complete successfully.