look for sort instead of zip during unit tests
On debian, sort is part of coreutils.  I think it ships by default on
any modern MacOS system.  And SORT.EXE is available on DOS and Windows
too (https://www.computerhope.com/sorthlp.htm).  zip, on the other
hand, isn't installed by default on debian, and i don't know that it's
available on Windows.

Making this change should still work to exercise the test, but without
making zip a dependency for the test suite itself.