lang/ca/help: use HTML entities consistently
A few stray characters with the high-bit set made it into these files.

This means that the file is most likely ISO-8859-1.  But like any flat
8-bit text encoding, it's not immediately obvious which 8-bit flat
encoding is the right one.  it could also be windows codepage 1252,
ISO-8859-15, etc.  Without an HTML charset hint in the <head> like:

    <meta http-equiv="content-type" content="text/html; charset=iso-8859-1"/>

the file could be easily misinterpreted.

The three sensible and consistent approaches are:

 a) include an HTML charset hint indicating the 8-bit encoding, and
    convert all the HTML &-escaped entities to that charset,

 b) convert to UTF-8 (and also include an HTML charset hint)

 c) use only 7-bit clean ASCII and HTML entities.

This change implements choice (c).