return valid signing and encryption subkeys
If we don't return the valid signing-capable subkeys, it's possible
that the message will be signed but be unverifiable, which is pretty
weird.

We don't need to include authentication-capable subkeys, since they
don't interact with e-mail.

Possibly including multiple encryption-capable subkeys is useful for
peers who don't have all the various encrpytion-capable algorithms
implemented.

This change may make the Autocrypt headers larger than they were
before, but for sensible OpenPGP certificates, it'll still be small
enough to avoid raising problems on modern MTAs.