Use distinct GnuPG homedirs for each test
As thunderbird has become more capable of running tests
asynchronously, having multiple tests which use the same GnuPG homedir
can cause problems with the test suite.

In particular, if test A is launched, and completes successfully while
test B is being launched, then withTestGpgHome from test A might
delete the homedir while test B has already begun.

If this happens, it can result in some weird race conditions.

This fix doesn't guarantee that the homedir is distinct (perhaps we
should change initializeGpgHome() to createUnique instead of just
create to make the directory?), but it makes a collision very
improbable.