Drop xvfb everywhere.
Enigmail should be able to rely on thunderbird's --headless argument
to avoid the need for an X11 server, at least as of thunderbird 67 (i
couldn't get it to work right for thunderbird 60).

Pull this out of all the CI invocations.

(i note while doing this that provision-travis.sh also refers to
jsunit-0.1.4, which is no longer valid.  if this is not being
maintained, it should be deleted from the repo, as it only adds to
confusion)